Create Secret number
Enter a number. If the number is equal to the secret number, you are free to go. Otherwise, you will be stuck in a while loop where your number will be either added by 1 or subtracted by 1 depending on if it is greater than or less than the secret number. Once your number is equal to the secret number, then the game will end.